backend/django/hestami_ai_project/requirements.txt
Django==5.1.1
psycopg2==2.9.9
pillow==10.4.0
dj-rest-auth==6.0.0
django-allauth==65.0.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.1.0
daphne==4.1.2
requests==2.32.3
whitenoise==6.8.2
celery>=5.3.0
redis>=4.5.0
django-celery-results>=2.5.0
moviepy>=2.0.0
python-json-logger==2.0.7
python-magic==0.4.27
#python-magic-bin==0.4.14 #Dev Windows-only requirement; does not work in Linux Docker --> Dockerfile has libmagic installed
numpy<2.0.0
opencv-python-headless==4.8.0.74
drf-spectacular==0.28.0
clamd==1.0.2
squareup==40.0.0.220250123
django-money==3.5.3
temporalio==1.9.0